requirements.txt
streamlit>=1.31.0
pandas>=2.1.0
plotly>=5.18.0
openpyxl>=3.1.0
scikit-learn>=1.5.0
anthropic>=0.80.0
python-dotenv>=1.0.0
streamlit-shadcn-ui>=0.1.19
